WP-CFM 1.0.3
git-svn-id: https://plugins.svn.wordpress.org/wp-cfm/trunk@925467 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/class-readwrite.php
@@ -32,22 +32,29 @@ class WPCFM_Readwrite
 
     /**
      * Move the file bundle to DB
-     * Pull is difficult; we need to figure out how to properly
-     * import each setting into the database
+     * @param string $bundle_name The bundle name (or "all")
      */
     function pull_bundle( $bundle_name ) {
-        $data = $this->read_file( $bundle_name );
-        $this->write_db( $bundle_name, $data );
+        $bundles = ( 'all' == $bundle_name ) ? $this->helper->get_bundle_names() : array( $bundle_name );
+
+        foreach ( $bundles as $bundle_name ) {
+            $data = $this->read_file( $bundle_name );
+            $this->write_db( $bundle_name, $data );
+        }
     }
 
 
     /**
      * Move the DB bundle to file
-     * Push is easy; we simply write data to file
+     * @param string $bundle_name The bundle name (or "all")
      */
     function push_bundle( $bundle_name ) {
-        $data = $this->read_db( $bundle_name );
-        $this->write_file( $bundle_name, json_encode( $data, JSON_PRETTY_PRINT ) );
+        $bundles = ( 'all' == $bundle_name ) ? $this->helper->get_bundle_names() : array( $bundle_name );
+
+        foreach ( $bundles as $bundle_name ) {
+            $data = $this->read_db( $bundle_name );
+            $this->write_file( $bundle_name, json_encode( $data, JSON_PRETTY_PRINT ) );
+        }
     }
 
 
@@ -98,7 +105,7 @@ class WPCFM_Readwrite
      * @return array
      */
     function read_file( $bundle_name ) {
-        if ( file_exists( "$this->folder/$bundle_name.json" ) ) {
+        if ( is_readable( "$this->folder/$bundle_name.json" ) ) {
             $contents = file_get_contents( "$this->folder/$bundle_name.json" );
             return json_decode( $contents, true );
         }
@@ -110,7 +117,16 @@ class WPCFM_Readwrite
      * Write the bundle to file
      */
     function write_file( $bundle_name, $data ) {
-        return file_put_contents( "$this->folder/$bundle_name.json", $data );
+        $filename = "$this->folder/$bundle_name.json";
+        if ( file_exists( $filename ) ) {
+            if ( is_writable( $filename ) ) {
+                return file_put_contents( $filename, $data );
+            }
+        }
+        elseif ( is_writable( $this->folder ) ) {
+            return file_put_contents( $filename, $data );
+        }
+        return false;
     }
 
 
@@ -118,7 +134,10 @@ class WPCFM_Readwrite
      * Delete a bundle file
      */
     function delete_file( $bundle_name ) {
-        return unlink( "$this->folder/$bundle_name.json" );
+        if ( is_writable( "$this->folder/$bundle_name.json" ) ) {
+            return unlink( "$this->folder/$bundle_name.json" );
+        }
+        return false;
     }
 
 
@@ -175,6 +194,9 @@ class WPCFM_Readwrite
             if ( ! empty( $db_data[ $key ]['callback'] ) ) {
                 $callback = $db_data[ $key ]['callback'];
             }
+
+            // Allow for callback override
+            $callback = apply_filters( 'wpcfm_pull_callback', $callback, $callback_params );
 
             if ( is_callable( $callback ) ) {
                 if ( is_array( $callback ) ) {
